requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
environs==9.5.0
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
isort==5.13.2
joblib==1.4.0
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.33
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-vector-stores-milvus==0.1.10
llama-parse==0.4.2
llamaindex-py-client==0.1.19
marshmallow==3.21.1
milvus==2.3.5
minio==7.2.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.23.6
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.1
pyarrow==16.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pymilvus==2.4.1
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
striprtf==0.0.26
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
